python: Move dazl.model.core.ContractMatch to dazl.query.
