+++ python/dazl/query/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2017-2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+__all__ = ["ContractMatch", "is_match"]
+
+from .query import ContractMatch, is_match
